# api-spec.yaml
openapi: 3.0.0
info:
  title: Game Hub Manager API
  description: Detailed API documentation for the Game Hub Manager Interface.
  version: 1.0.0
servers:
  - url: /api
    description: Local API server
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session_token
paths:
...
  /auth/login:
    post:
      tags: [Auth]
      summary: Login
      description: Authenticate user with email and password. Sets a session cookie.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email: { type: string }
                password: { type: string }
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  user: { $ref: '#/components/schemas/User' }
        400: { description: Email or password required }
        401: { description: Invalid credentials }
        403: { description: Account disabled }
...
  /games/list:
    get:
      tags: [Games]
      summary: List games
      description: Returns a list of games based on user roles and filters.
      parameters:
        - name: status
          in: query
          schema: { type: string }
        - name: ownerId
          in: query
          schema: { type: string }
        - name: subject
          in: query
          schema: { type: string }
        - name: grade
          in: query
          schema: { type: string }
        - name: isDeleted
          in: query
          schema: { type: boolean }
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  games:
                    type: array
                    items: { $ref: '#/components/schemas/GameWithVersion' }
        401: { description: Unauthorized }
  /games/create:
    post:
      tags: [Games]
      summary: Create game
      description: Creates a new game and an initial version (1.0.0).
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/GameCreateInput' }
      responses:
        201:
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  game: { $ref: '#/components/schemas/Game' }
                  version: { $ref: '#/components/schemas/GameVersion' }
        400: { description: Validation error or Game ID exists }
        403: { description: Forbidden }
...
  /games/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string }
    get:
      tags: [Games]
      summary: Get game details
      responses:
        200: { description: OK }
        404: { description: Not found }
...
components:
  schemas:
    User:
      type: object
      properties:
        id: { type: string }
        email: { type: string }
        name: { type: string }
        roles: { type: array, items: { type: string } }
        avatar: { type: string }
        isActive: { type: boolean }
    Game:
      type: object
      properties:
        _id: { type: string }
        gameId: { type: string }
        title: { type: string }
        description: { type: string }
        ownerId: { type: string }
        subject: { type: string }
        grade: { type: string }
        status: { type: string, description: "Deprecated" }
        latestVersionId: { type: string }
        liveVersionId: { type: string }
        isDeleted: { type: boolean }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    GameWithVersion:
      allOf:
        - $ref: '#/components/schemas/Game'
        - type: object
          properties:
            latestVersion:
              type: object
              properties:
                _id: { type: string }
                version: { type: string }
                status: { type: string }
                submittedAt: { type: string, format: date-time }
            liveVersion:
              type: object
              properties:
                _id: { type: string }
                version: { type: string }
    GameVersion:
      type: object
      properties:
        _id: { type: string }
        gameId: { type: string }
        version: { type: string }
        status: { type: string, enum: [draft, uploaded, qc_passed, qc_failed, approved, rejected, published] }
        storagePath: { type: string }
        entryFile: { type: string }
        submittedBy: { type: string }
        submittedAt: { type: string, format: date-time }
        selfQAChecklist: { $ref: '#/components/schemas/SelfQAChecklist' }
    SelfQAChecklist:
      type: object
      properties:
        testedDevices: { type: boolean }
        testedAudio: { type: boolean }
        gameplayComplete: { type: boolean }
        contentVerified: { type: boolean }
        note: { type: string }
    GameCreateInput:
      type: object
      required: [title, gameId]
      properties:
        title: { type: string }
        gameId: { type: string }
        subject: { type: string }
        grade: { type: string }
        unit: { type: string }
        gameType: { type: string }
        priority: { type: string }
        description: { type: string }
...